requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
bettercam==1.0.0
huggingface-hub==0.22.2
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
safetensors==0.4.3
scipy==1.13.0
tk==0.1.0
tqdm==4.66.2
urllib3==2.2.1
websocket-client==1.7.0
winclip32==0.6.0
